requirements.txt
requests
lxml
pandas
scikit-learn
python-dotenv
flask
gunicorn
